task week five, regulator circuit with DC
here i add two circuits, the first one by AC transferring and regulating to become a DC then distribution in three voltage
the second is using DC current and distributing it into three voltages, to get the specific voltage from a source having 35 volts are using regulators IC, so LM7805 for 5v, LM7809 for9v and LM7012 for 12v. The capacitor charges and discharges during a cycle to give a steady DC voltage as the output, so These capacitors act as a filtering capacitor that bypasses all the AC through them to the ground as is shown in the circuit that depends on the AC source. At the output, the mean DC voltage left is smoother and ripple free.